Move Future and Either to types folder
