+++ src/domain/types/Future.ts
