+++ src/domain/usecases/GetInstanceVersionUseCase.ts
@@ -1,5 +1,5 @@
 import { UseCase } from "../../CompositionRoot";
-import { FutureData } from "../entities/Future";
+import { FutureData } from "../types/Future";
 import { InstanceRepository } from "../repositories/InstanceRepository";
 
 export class GetInstanceVersionUseCase implements UseCase {
